Ignore any query params without corresponding formatter functions
Considering that tokens or numerous other things
may be included in the req.query object, it is
probably more prudent to ignore them by default,
rather than include them, and having some strange
values included in a query.

If all query params were included by default, the
client could mess with the query string before it
is sent to the backend, meaning we could end up
with lots of unexpected values on our prisma query,
with unexpected consequences.

So for that reason, query params without a defined
formatter will be ignored by default instead of
included on the where object.