Do not process query values in any way unless defined in a formatter
Previously if a query value could be cast to a Number,
it was. However this may not be the desired behaviour
in some cases, so this functionality has been removed.
The user can use a processor function to turn the query
string value into a number or date or whatever they need.