Allow usage of different path separators to access cols of foreign relations
With the new customOptions parameter on urlQueryToPrisma,
the user can define a pathSeparator. This defaults to a
dot. This object gets passed to every formatter as they
are called.

In the future I will try to simplify the parameters for
end users, e.g. move 'querySource' into the customOptions
object instead of being its own parameter.