formatters.where can now match against column with a different name/path to the url param
formatterOptions parameter with tableColName property
can be used to check a url param against a table col
with a name different from it.

Previously, the url param had to match the table col
name.

Like the urlQueryToPrisma middleware itself, I think
the parameters to the formatter functions need simplifying.
Currently we have filterType, filterOptions, valueProcessor
and now formatterOptions, all of which are optional.

An customOptions object that can be merged with defaultOptions
would be better, with the user having to repeatedly set default
options (e.g. in order to set the 4th parameter, you have to set
the 1st - 3rd parameters, which you may not be interested in).