Use customOptions parameter instead of individual params
Previously, the user would have to set all params
individually even if they only wanted to change
the 3rd or 4th parameter from its default value.
Especially annoying was having to define a value
processor function just to be able to set the
4th parameter.

Now, the user provided params on the options object
will merge with the defaults, so they only have to
set the parameters they want to change from default.